move more types and getWranglerTmpDir into workers-utils
Moves Entry, EphemeralDirectory, AssetsOptions, LegacyAssetPaths types
