+++ packages/workers-utils/package.json
@@ -50,12 +50,14 @@
 		"@cloudflare/workflows-shared": "workspace:*",
 		"@types/command-exists": "^1.2.0",
 		"@types/node": "catalog:default",
+		"@types/signal-exit": "^3.0.1",
 		"@vitest/ui": "catalog:default",
 		"cloudflare": "^5.2.0",
 		"command-exists": "^1.2.9",
 		"concurrently": "^8.2.2",
 		"empathic": "^2.0.0",
 		"jsonc-parser": "catalog:default",
+		"signal-exit": "catalog:default",
 		"smol-toml": "catalog:default",
 		"ts-dedent": "^2.2.0",
 		"tsdown": "^0.15.9",
