+++ packages/workers-utils/src/types.ts
@@ -42,8 +42,9 @@ import type {
 	CfVpcService,
 	CfWorkerLoader,
 	CfWorkflow,
+	CfScriptFormat,
 } from "./worker";
-import type { AssetConfig } from "@cloudflare/workers-shared";
+import type { AssetConfig, RouterConfig } from "@cloudflare/workers-shared";
 
 export type Json =
 	| string
@@ -203,6 +204,40 @@ export type AssetConfigMetadata = {
 	_headers?: string;
 };
 
+export type AssetsOptions = {
+	directory: string;
+	binding?: string;
+	routerConfig: RouterConfig;
+	assetConfig: AssetConfig;
+	_redirects?: string;
+	_headers?: string;
+	run_worker_first?: boolean | string[];
+};
+
+/**
+ * Information about the assets that should be uploaded
+ */
+export interface LegacyAssetPaths {
+	/**
+	 * Absolute path to the root of the project.
+	 *
+	 * This is the directory containing wrangler.toml or cwd if no config.
+	 */
+	baseDirectory: string;
+	/**
+	 * The path to the assets directory, relative to the `baseDirectory`.
+	 */
+	assetDirectory: string;
+	/**
+	 * An array of patterns that match files that should be uploaded.
+	 */
+	includePatterns: string[];
+	/**
+	 * An array of patterns that match files that should not be uploaded.
+	 */
+	excludePatterns: string[];
+}
+
 // for PUT /accounts/:accountId/workers/scripts/:scriptName
 type WorkerMetadataPut = {
 	/** The name of the entry point module. Only exists when the worker is in the ES module format */
@@ -354,3 +389,28 @@ export type Binding =
 	| ({ type: `unsafe_${string}` } & Omit<CfUnsafeBinding, "name" | "type">)
 	| { type: "assets" }
 	| { type: "inherit" };
+
+/**
+ * An entry point for the Worker.
+ *
+ * It consists not just of a `file`, but also of a `directory` that is used to resolve relative paths.
+ */
+export type Entry = {
+	/** A worker's entrypoint */
+	file: string;
+	/** A worker's directory. Usually where the Wrangler configuration file is located */
+	projectRoot: string;
+	/** The path to the config file, if it exists. */
+	configPath: string | undefined;
+	/** Is this a module worker or a service worker? */
+	format: CfScriptFormat;
+	/** The directory that contains all of a `--no-bundle` worker's modules. Usually `${directory}/src`. Defaults to path.dirname(file) */
+	moduleRoot: string;
+	/**
+	 * A worker's name
+	 */
+	name?: string | undefined;
+
+	/** Export from a Worker's entrypoint */
+	exports: string[];
+};
