+++ packages/wrangler/package.json
@@ -151,7 +151,7 @@
 		"resolve": "^1.22.8",
 		"semiver": "^1.1.0",
 		"shell-quote": "^1.8.1",
-		"signal-exit": "^3.0.7",
+		"signal-exit": "catalog:default",
 		"smol-toml": "catalog:default",
 		"source-map": "^0.6.1",
 		"supports-color": "^9.2.2",
