+++ packages/wrangler/src/api/startDevWorker/types.ts
@@ -1,10 +1,10 @@
-import type { AssetsOptions } from "../../assets";
 import type { CfAccount } from "../../dev/create-worker-preview";
 import type { EsbuildBundle } from "../../dev/use-esbuild";
 import type { ConfigController } from "./ConfigController";
 import type { DevEnv } from "./DevEnv";
 import type { ContainerNormalizedConfig } from "@cloudflare/containers-shared";
 import type {
+	AssetsOptions,
 	BinaryFile,
 	Binding,
 	CfModule,
