+++ packages/wrangler/src/check/commands.ts
@@ -4,7 +4,7 @@ import { readFile, writeFile } from "node:fs/promises";
 import path from "node:path";
 import { log } from "@cloudflare/cli-shared-helpers";
 import { spinnerWhile } from "@cloudflare/cli-shared-helpers/interactive";
-import { UserError } from "@cloudflare/workers-utils";
+import { getWranglerTmpDir, UserError } from "@cloudflare/workers-utils";
 import chalk from "chalk";
 import { Miniflare } from "miniflare";
 import { WebSocket } from "ws";
@@ -16,7 +16,6 @@ import {
 	ModuleTypeToRuleType,
 } from "../deployment-bundle/module-collection";
 import { logger } from "../logger";
-import { getWranglerTmpDir } from "../paths";
 import type { Config } from "@cloudflare/workers-utils";
 import type { ModuleDefinition } from "miniflare";
 import type { FormData, FormDataEntryValue } from "undici";
