+++ packages/wrangler/src/deployment-bundle/bundle.ts
@@ -3,12 +3,13 @@ import * as path from "node:path";
 import {
 	getBuildConditionsFromEnv,
 	getBuildPlatformFromEnv,
+	getWranglerTmpDir,
 	UserError,
 } from "@cloudflare/workers-utils";
 import chalk from "chalk";
 import * as esbuild from "esbuild";
 import { getFlag } from "../experimental-flags";
-import { getBasePath, getWranglerTmpDir } from "../paths";
+import { getBasePath } from "../paths";
 import { applyMiddlewareLoaderFacade } from "./apply-middleware";
 import {
 	isBuildFailure,
@@ -23,13 +24,13 @@ import { getNodeJSCompatPlugins } from "./esbuild-plugins/nodejs-plugins";
 import { writeAdditionalModules } from "./find-additional-modules";
 import { noopModuleCollector } from "./module-collection";
 import type { MiddlewareLoader } from "./apply-middleware";
-import type { Entry } from "./entry";
 import type { ModuleCollector } from "./module-collection";
 import type {
 	CfModule,
 	CfModuleType,
 	Config,
 	DurableObjectBindings,
+	Entry,
 	WorkflowBinding,
 } from "@cloudflare/workers-utils";
 import type { NodeJSCompatMode } from "miniflare";
