+++ packages/wrangler/src/dev/miniflare/index.ts
@@ -20,11 +20,10 @@ import { updateCheck } from "../../update-check";
 import { warnOrError } from "../../utils/print-bindings";
 import { getDurableObjectClassNameToUseSQLiteMap } from "../class-names-sqlite";
 import type { StartDevWorkerInput } from "../../api/startDevWorker/types";
-import type { AssetsOptions } from "../../assets";
 import type { LoggerLevel } from "../../logger";
-import type { LegacyAssetPaths } from "../../sites";
 import type { EsbuildBundle } from "../use-esbuild";
 import type {
+	AssetsOptions,
 	Binding,
 	CfD1Database,
 	CfDispatchNamespace,
@@ -37,6 +36,7 @@ import type {
 	CfWorkflow,
 	Config,
 	ContainerEngine,
+	LegacyAssetPaths,
 } from "@cloudflare/workers-utils";
 import type {
 	DOContainerOptions,
