+++ packages/wrangler/src/utils/friendly-validator-errors.ts
@@ -1,10 +1,9 @@
 import { writeFile } from "node:fs/promises";
 import path from "node:path";
-import { ParseError } from "@cloudflare/workers-utils";
+import { getWranglerTmpDir, ParseError } from "@cloudflare/workers-utils";
 import dedent from "ts-dedent";
 import { analyseBundle } from "../check/commands";
 import { logger } from "../logger";
-import { getWranglerTmpDir } from "../paths";
 import type { Metafile } from "esbuild";
 import type { FormData } from "undici";
